Run tests via pytest when pytest is a test-req (via pipfile)

+++ examples/refresh.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 
 import argparse
-import io
 import logging
 import os
 import subprocess
@@ -14,9 +13,7 @@ COMMANDS = 'explain entrypoints'.split()
 
 
 sys.path.insert(0, PPATH)
-
-
-import setupmeta
+import setupmeta            # noqa
 
 
 def run_command(path, command):
@@ -33,16 +30,21 @@ def run_command(path, command):
     return setupmeta.to_str(output) + setupmeta.to_str(error)
 
 
-def refresh_example(path):
+def refresh_example(path, dryrun):
     logging.info("Refreshing %s" % (os.path.basename(path)))
     setup_py = os.path.join(path, 'setup.py')
     if not os.path.isfile(setup_py):
         return
     expected = os.path.join(path, 'expected.txt')
+    output = ''
+    for command in COMMANDS:
+        output += "Replay: %s\n" % command
+        output += "%s\n" % run_command(path, command)
+    if dryrun:
+        print(output)
+        return
     with open(expected, 'w') as fh:
-        for command in COMMANDS:
-            fh.write("Replay: %s\n" % command)
-            fh.write("%s\n" % run_command(path, command))
+        fh.write(output)
 
 
 def main():
@@ -55,10 +57,18 @@ def main():
         action='store_true',
         help="Show debug info"
     )
+    parser.add_argument(
+        '-n', '--dryrun',
+        action='store_true',
+        help="Print output rather, don't update expected.txt"
+    )
     args = parser.parse_args()
 
     level = logging.DEBUG if args.debug else logging.INFO
-    logging.basicConfig(format="%(asctime)s %(levelname)s %(message)s", level=level)
+    logging.basicConfig(
+        format="%(asctime)s %(levelname)s %(message)s",
+        level=level
+    )
     logging.root.setLevel(level)
 
     examples = os.listdir(EXAMPLES)
@@ -66,7 +76,7 @@ def main():
         path = os.path.join(EXAMPLES, name)
         if not os.path.isdir(path):
             continue
-        refresh_example(path)
+        refresh_example(path, dryrun=args.dryrun)
 
 
 if __name__ == "__main__":
